Clean up table persistence
By default (in production), the new decorator will store data frames
to disk.  In debug mode, it lazy-loads the data from disk and
short-circuits calculation to save expensive query resources.

Responsibility for persistence is moved out of the query modules.

One module is split because we want to persist both of the tables it
generates.

Change-Id: I03e82f9cb72c63e0ab4df45d0ca8aead545f5fe6