Remove known buggy conflicts
There are two huge, unsolved mysteries accounting for up to 2/3 of
all conflicts.  Remove this data from the main sample because it may
have systematic biases which damage the rest of our analysis, for
example it may be unsolvable or always successfully resolved.  These
bugs need to be investigated on their own.

Change-Id: Ib2fcb1b63b5fdb0a36a59d258f0a7ccb60b24eb5